Fixes test to work with oci ref
